Unit test for encoding to string w/ checksum
