+++ test/unit/KeyTest.php
@@ -32,34 +32,4 @@ class KeyTest extends PHPUnit_Framework_TestCase
         Key::loadFromAsciiSafeString($str);
     }
 
-    /**
-     * @expectedException \Defuse\Crypto\Exception\BadFormatException
-     * @expectedExceptionMessage checksum doesn't match
-     */
-    public function testIncorrectChecksum()
-    {
-        $key                                     = Key::createNewRandomKey();
-        $str                                     = $key->saveToAsciiSafeString();
-        $str[2*Encoding::SERIALIZE_HEADER_BYTES + 0] = 'f';
-        $str[2*Encoding::SERIALIZE_HEADER_BYTES + 1] = 'f';
-        $str[2*Encoding::SERIALIZE_HEADER_BYTES + 3] = 'f';
-        $str[2*Encoding::SERIALIZE_HEADER_BYTES + 4] = 'f';
-        $str[2*Encoding::SERIALIZE_HEADER_BYTES + 5] = 'f';
-        $str[2*Encoding::SERIALIZE_HEADER_BYTES + 6] = 'f';
-        $str[2*Encoding::SERIALIZE_HEADER_BYTES + 7] = 'f';
-        $str[2*Encoding::SERIALIZE_HEADER_BYTES + 8] = 'f';
-        Key::loadFromAsciiSafeString($str);
-    }
-
-    /**
-     * @expectedException \Defuse\Crypto\Exception\BadFormatException
-     * @expectedExceptionMessage not a hex string
-     */
-    public function testBadHexEncoding()
-    {
-        $key    = Key::createNewRandomKey();
-        $str    = $key->saveToAsciiSafeString();
-        $str[0] = 'Z';
-        Key::loadFromAsciiSafeString($str);
-    }
 }
